podziel na moduly
