+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/logic.ts
@@ -1,0 +1,11 @@
+import {User} from './data'
+
+function hasAddress(user: User): boolean {
+    return Boolean(user.address)
+  }
+  
+function hasGivenAge(requiredAge: number): any {
+    return (user: User): boolean => user.age >= requiredAge
+}
+
+export {hasAddress, hasGivenAge}
