+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/main.ts
@@ -1,38 +1,5 @@
-/**
- * Zasady co do pliku:
- *
- * Możesz dowolnie modyfikować zawartość tego pliku,
- * całość programu musi jednak działać tak jak do tej pory !
- *
- * */
-
- interface User {
-  name: string,
-  age: number,
-  email: string,
-  address: {
-    street: string,
-    no: number
-  },
-}
-
-const user: User = {
-  name: 'Andy',
-  age: 30,
-  email: 'andy@mail-me-tommorow.com',
-  address: {
-    street: 'Strange Alley',
-    no: 23,
-  },
-}
-
-function hasAddress(user: User): boolean {
-  return Boolean(user.address)
-}
-
-function hasGivenAge(requiredAge: number): any {
-  return (user: User): boolean => user.age >= requiredAge
-}
+import {user} from './data'
+import {hasAddress, hasGivenAge} from './logic'
 
 const isAdult = hasGivenAge(18)
